% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoring_variables.R
\name{add_factor_contrasts}
\alias{add_factor_contrasts}
\title{Add treatment contrasts to a factor}
\usage{
add_factor_contrasts(vect, n = levels(vect), ...)
}
\arguments{
\item{vect}{A factor vector to which contrasts will be added.}

\item{n}{A vector of levels for the factor. Default is the levels of \code{vect}.}

\item{...}{Additional arguments passed to \code{contr.treatment}.}
}
\value{
A factor vector with treatment contrasts added.
}
\description{
Add treatment contrasts to a factor
}
