% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphantasiaReasoningViie-package.R
\docType{package}
\name{aphantasiaReasoningViie-package}
\alias{aphantasiaReasoningViie}
\alias{aphantasiaReasoningViie-package}
\title{aphantasiaReasoningViie: Data Analysis for the EMC Lab.'s Study on Aphantasia and Reasoning}
\description{
This package contains all the code to reproduce the analyses associated with the 2025 online study on aphantasia and reasoning.
}
\author{
\strong{Maintainer}: Maël Delem \email{mael.delem@pm.me} (\href{https://orcid.org/0009-0005-8518-1991}{ORCID})

}
\keyword{internal}
