% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tools.R
\name{build_formula}
\alias{build_formula}
\title{Build a formula based on the common model for accuracy and RT modelling}
\usage{
build_formula(vd, grouping)
}
\arguments{
\item{vd}{Variable of interest, either "accuracy" or "rt_total".}

\item{grouping}{Grouping variable, typically "Group" or "cluster".}
}
\value{
A formula object for the model.
}
\description{
Build a formula based on the common model for accuracy and RT modelling
}
\examples{
build_formula("accuracy", "group")
}
