% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_nieq_scores}
\alias{compute_nieq_scores}
\title{Compute NIEQ scores by combining the frequency and proportion items of each
subscale}
\usage{
compute_nieq_scores(df)
}
\arguments{
\item{df}{A data frame containing the NIEQ data with columns for frequency
and proportion of each subscale.}
}
\value{
A data frame with new columns for NIEQ scores, calculated as the
mean of the frequency and proportion items for each subscale.
}
\description{
Compute NIEQ scores by combining the frequency and proportion items of each
subscale
}
\examples{
df <- compute_nieq_scores(survey_data)
df |>
  dplyr::filter(nieq_is_complete == TRUE) |>
  dplyr::select("id", tidyselect::contains("nieq_")) |>
  head()
}
