% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{describe_survey_data}
\alias{describe_survey_data}
\title{Helper function to quickly describe the cleaned survey data}
\usage{
describe_survey_data(df, grouping = NULL)
}
\arguments{
\item{df}{A data frame containing the survey data columns related to the
questionnaires. Note that the functions computes means for the combined
NIEQ scores, so the data frame must have gone through
\code{\link[=compute_nieq_scores]{compute_nieq_scores()}}.}
}
\value{
A data frame summarising the survey data grouped by VVIQ group,
including sample size, gender distribution, mean scores for VVIQ, OSIVQ
sub-scales, Raven matrices and NIEQ sub-scales, along with the number of
participants who completed the NIEQ questionnaire (which was part of another
online experiment, so completely optional).
}
\description{
Helper function to quickly describe the cleaned survey data
}
\examples{
# The filtering function removes participants with bad or incomplete data,
# notably those who did not complete certain questionnaires.
survey_data |>
 filter_manually_identified_ids() |>
 compute_nieq_scores() |>
 describe_survey_data()

}
\keyword{internal}
