% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_utils.R
\name{extract_emm_confint}
\alias{extract_emm_confint}
\title{Extract and format the confidence interval of an emmeans object}
\usage{
extract_emm_confint(emm_object)
}
\arguments{
\item{emm_object}{An emm_grid object from which to extract the confidence
interval.}
}
\value{
A data frame with the confidence interval formatted as a string.
}
\description{
Extract and format the confidence interval of an emmeans object
}
\keyword{internal}
