% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoring_variables.R
\name{factor_categories}
\alias{factor_categories}
\title{Convert the \code{category} column to a factor with contrasts}
\usage{
factor_categories(df, n = c("_control", "_spatial", "_visual"), base = 3, ...)
}
\arguments{
\item{df}{A data frame containing a \code{category} column.}

\item{n}{A vector of levels for the factor. Default is
\code{c("_control", "_spatial", "_visual")}.}

\item{base}{The base level for the contrasts. Default is 3 for the visual
category, as our hypothesis called for planned comparisons of the visual
category with the other two. This argument can be used to set a different
base level if desired.}

\item{...}{Additional arguments passed to \code{\link[=add_factor_contrasts]{add_factor_contrasts()}}.}
}
\value{
A data frame with the \code{category} column converted to a factor with
contrasts added.
}
\description{
Convert the \code{category} column to a factor with contrasts
}
\examples{
df <- factor_categories(experiment_data)
contrasts(df$category)
}
