% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoring_variables.R
\name{factor_chr_vars}
\alias{factor_chr_vars}
\title{Convert all character variables in a data frame to factors}
\usage{
factor_chr_vars(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with all character variables converted to factors.
}
\description{
Convert all character variables in a data frame to factors
}
