% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoring_variables.R
\name{factor_groups}
\alias{factor_groups}
\title{Convert the \code{group} column to a factor with the desired VVIQ classification}
\usage{
factor_groups(df, n_groups = 2, contrast_base = NULL, ...)
}
\arguments{
\item{df}{A data frame containing a \code{group} column with VVIQ groups
among "Aphantasia", "Hypophantasia", "Typical" and "Hyperphantasia".}

\item{n_groups}{An integer indicating the number of groups to create, which
reflect specific classifications of VVIQ scores: two groups mean grouping
Aphants and Hypophants on one end and Typical and Hyperphants on the other,
three groups mean separating Aphants, Hypophants and Typical but removing
Hyperphants (that have been shown to have distinct characteristics), and four
groups mean separating and keeping all four groups. Default is 2.}

\item{contrast_base}{Optional. An integer indicating the base level for the
contrasts. By default, the base level is set to the Typical group for all
classifications. This argument allows you to set a different base level
(i.e., a different group to compare against for planned comparisons).}

\item{...}{Additional arguments passed to \code{\link[=add_factor_contrasts]{add_factor_contrasts()}}.}
}
\value{
A data frame with the \code{group} column converted to a factor with
contrasts added, and the levels set according to the number of groups.
}
\description{
Convert the \code{group} column to a factor with the desired VVIQ classification
}
\examples{
df <- factor_groups(survey_data, n_groups = 2)
contrasts(df$group)

df <- factor_groups(survey_data, n_groups = 3)
contrasts(df$group)

df <- factor_groups(survey_data, n_groups = 4)
contrasts(df$group)

# Choosing Aphantasia as the reference level for the contrasts
# (i.e., for planned comparisons)
df <- factor_groups(survey_data, n_groups = 4, contrast_base = 1)
contrasts(df$group)
}
