% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoring_variables.R
\name{factor_strategies}
\alias{factor_strategies}
\title{Convert strategy columns to factors with optionally ordered levels}
\usage{
factor_strategies(df, ordered = TRUE)
}
\arguments{
\item{df}{A data frame containing strategy columns to be converted to
factors.}

\item{ordered}{A logical indicating whether the factor levels should be
ordered.}
}
\value{
A data frame with strategy columns converted to factors, with levels
ranging from 1 to 5 for ordered factors, or as character strings for
unordered factors.
}
\description{
Convert strategy columns to factors with optionally ordered levels
}
\examples{
# The filtering function removes participants with bad or incomplete data,
# notably those who did not complete the strategy questionnaire.
df <- survey_data |>
  filter_manually_identified_ids() |>
  factor_strategies(ordered = TRUE)
levels(df$visual_strat)
contrasts(df$visual_strat)

df <- survey_data |>
  filter_manually_identified_ids() |>
  factor_strategies(ordered = FALSE)
levels(df$visual_strat)
contrasts(df$visual_strat)
}
