% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_participants.R
\name{filter_manually_identified_ids}
\alias{filter_manually_identified_ids}
\title{Filter manually identified participant based on various criteria}
\usage{
filter_manually_identified_ids(
  df,
  exclude_no_vviq = TRUE,
  exclude_no_osivq = TRUE,
  exclude_no_raven = TRUE,
  exclude_cheated = TRUE,
  exclude_distracted = TRUE,
  exclude_treatment = FALSE,
  exclude_adhd = FALSE,
  exclude_asd = FALSE,
  exclude_dyslexia = FALSE,
  exclude_other = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{A data frame containing participant information.}

\item{exclude_no_vviq}{Logical, whether to exclude participants without VVIQ.}

\item{exclude_no_osivq}{Logical, whether to exclude participants without
OSIVQ.}

\item{exclude_no_raven}{Logical, whether to exclude participants without
Raven.}

\item{exclude_cheated}{Logical, whether to exclude participants who have
cheated (based on self-report).}

\item{exclude_distracted}{Logical, whether to exclude participants who have
been distracted (based on self-report).}

\item{exclude_treatment}{Logical, whether to exclude participants who have
a treatment for a neurological or psychiatric disorder.}

\item{exclude_adhd}{Logical, whether to exclude participants who have ADHD.}

\item{exclude_asd}{Logical, whether to exclude participants who have ASD.}

\item{exclude_dyslexia}{Logical, whether to exclude participants who have
dyslexia.}

\item{exclude_other}{Logical, whether to exclude participants who have
other neurological troubles.}

\item{verbose}{Logical, whether to print the number of participants excluded
based on the criteria.}
}
\value{
A filtered data frame with participants who meet the specified
criteria.
}
\description{
Filter manually identified participant based on various criteria
}
\examples{
df <- filter_manually_identified_ids(survey_data)

# We might want to try being less strict and keeping people who did not
# complete the OSIVQ or Raven, e.g. if we're only interested in the VVIQ
df <-
 filter_manually_identified_ids(
  survey_data,
  exclude_no_osivq = FALSE,
  exclude_no_raven = FALSE
 )
# Alternatively we could study a sample without any psychiatric/neurological
# troubles, etc.
}
