% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_participants.R
\name{filter_random_accuracy_ids}
\alias{filter_random_accuracy_ids}
\title{Filter participants with below random accuracy}
\usage{
filter_random_accuracy_ids(df, verbose = TRUE)
}
\arguments{
\item{df}{A data frame containing participant responses and accuracy.}

\item{verbose}{Logical, whether to print the number of participants excluded
based on accuracy.}
}
\value{
A filtered data frame with participants who have an accuracy above
50\% across all trials.
}
\description{
Filter participants with below random accuracy
}
\examples{
df <- filter_random_accuracy_ids(experiment_data)
}
