% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_participants.R
\name{filter_suspicious_rt_ids}
\alias{filter_suspicious_rt_ids}
\title{Filter participants with suspicious median RTs}
\usage{
filter_suspicious_rt_ids(df, sd_mult = 2, verbose = TRUE)
}
\arguments{
\item{df}{A data frame containing participant responses and median RTs.}

\item{sd_mult}{A numeric value indicating how many standard deviations to use
for identifying suspicious median RTs. The default is 2, which means that
median RTs that are more than 2 standard deviations away from the mean
will be considered suspicious.}
}
\value{
A filtered data frame with participants who have median RTs within
the specified range.
}
\description{
Filter participants with suspicious median RTs
}
