% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_trials.R
\name{filter_trials_on_rt}
\alias{filter_trials_on_rt}
\title{Filter outlier trials based on mean response time per participant}
\usage{
filter_trials_on_rt(
  df,
  mult = 2,
  min_rt = 0.6,
  max_rt = 30,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{df}{A data frame containing trial data with columns for id, Group,
trial_number, category, premise_1_rt, premise_2_rt, premise_3_rt,
conclusion_rt and accuracy.}

\item{mult}{A multiplier for the standard deviation to set the threshold for
outlier detection. Default is 2.}

\item{min_rt}{Minimum response time in seconds for trials to be considered
valid. Default is 0.6 seconds.}

\item{max_rt}{Maximum response time in seconds for trials to be considered
valid. Default is 30 seconds.}

\item{verbose}{Logical. Whether to print a summary of the filtering
process. Default is FALSE.}

\item{...}{Additional arguments passed to the function. Unused.}
}
\value{
A filtered data frame containing only valid trials, with outliers
removed based on the specified thresholds.
}
\description{
Filter outlier trials based on mean response time per participant
}
\examples{
df_rt <-
  get_clean_data()$df_expe |>
  filter_trials_on_rt()
}
