% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tools.R
\name{fit_clm}
\alias{fit_clm}
\title{Fit a cumulative link model (CLM) using the ordinal package}
\usage{
fit_clm(formula, data, link = "probit")
}
\arguments{
\item{formula}{A formula specifying the model to fit.}

\item{data}{A data frame containing the data to fit the model.}

\item{link}{A string specifying the link function to use. Default is
"probit".}
}
\value{
A fitted clm object from the ordinal package.
}
\description{
Fit a cumulative link model (CLM) using the ordinal package
}
\examples{
df_strats <- get_clean_data()$df_survey |> pivot_strategies_longer()

model <- fit_clm(score ~ group_2 * strategy, df_strats)
report_contrast(model, ~ group_2 | strategy)
}
