% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_clean_data}
\alias{get_clean_data}
\title{Wrapper function to get clean "analysis-ready" data}
\usage{
get_clean_data(
  type = "all",
  n_groups = 2,
  exclude_no_vviq = TRUE,
  exclude_no_osivq = TRUE,
  exclude_no_raven = TRUE,
  exclude_cheated = TRUE,
  exclude_distracted = TRUE,
  exclude_treatment = FALSE,
  exclude_adhd = FALSE,
  exclude_asd = FALSE,
  exclude_dyslexia = FALSE,
  exclude_other = FALSE,
  sd_mult = 2.25,
  verbose = FALSE
)
}
\arguments{
\item{type}{The type of data to return: "experiment", "survey", or "all".}

\item{n_groups}{The number of groups to factor in the data. Must be 2, 3 or
4. 2 divides the sample into Aphants and Typical imagers using the 32 VVIQ
criterio, 3 divides the sample into Aphants (VVIQ = 16), Hypophants
(VVIQ < 32) and Typical imagers, and 4 also isolates Hyperphants with
VVIQ > 75.}

\item{exclude_no_vviq}{Logical, whether to exclude participants without VVIQ.}

\item{exclude_no_osivq}{Logical, whether to exclude participants without
OSIVQ.}

\item{exclude_no_raven}{Logical, whether to exclude participants without
Raven.}

\item{exclude_cheated}{Logical, whether to exclude participants who have
cheated (based on self-report).}

\item{exclude_distracted}{Logical, whether to exclude participants who have
been distracted (based on self-report).}

\item{exclude_treatment}{Logical, whether to exclude participants who have
a treatment for a neurological or psychiatric disorder.}

\item{exclude_adhd}{Logical, whether to exclude participants who have ADHD.}

\item{exclude_asd}{Logical, whether to exclude participants who have ASD.}

\item{exclude_dyslexia}{Logical, whether to exclude participants who have
dyslexia.}

\item{exclude_other}{Logical, whether to exclude participants who have
other neurological troubles.}

\item{sd_mult}{A numeric value indicating how many standard deviations to use
for identifying suspicious median RTs. The default is 2.25, which means that
median RTs that are more than 2.25 standard deviations inferior to the mean
are considered suspiciously fast and potential "spamming".}

\item{verbose}{A logical value indicating whether to print verbose messages
about the filtering process. Default is \code{FALSE}.}
}
\value{
A cleaned data frame or a list of cleaned data frames, depending on the
\code{type} parameter:
\itemize{
\item If \code{type} is "experiment", returns the cleaned experiment data frame.
\item If \code{type} is "survey", returns the cleaned survey data frame.
\item If \code{type} is "all", returns a list containing both cleaned data frames.
}
}
\description{
This function retrieves and cleans the data for the experiment and survey. It
uses several helper functions to filter and format the data, including
\code{\link[=filter_random_accuracy_ids]{filter_random_accuracy_ids()}}, \code{\link[=filter_manually_identified_ids]{filter_manually_identified_ids()}},
\code{\link[=filter_suspicious_rt_ids]{filter_suspicious_rt_ids()}}, \code{\link[=factor_categories]{factor_categories()}}, \code{\link[=factor_groups]{factor_groups()}},
\code{\link[=factor_chr_vars]{factor_chr_vars()}}, \code{\link[=factor_strategies]{factor_strategies()}}, and \code{\link[=compute_nieq_scores]{compute_nieq_scores()}}.
The cleaned data is returned as a list containing two data frames:
\code{df_expe} and \code{df_survey}. The \code{df_expe} data frame contains the cleaned
experiment data, while the \code{df_survey} data frame contains the cleaned
survey data.
}
\examples{
clean_data <- get_clean_data(type = "all", verbose = TRUE)
colnames(get_clean_data(type = "experiment"))
colnames(get_clean_data(type = "survey"))
}
