% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_clustered_data}
\alias{get_clustered_data}
\title{Wrapper function to get clean and clustered data immediately}
\usage{
get_clustered_data(
  type = "all",
  names = c("Spatialiser", "Visualiser", "Verbaliser"),
  levels = c("Visualiser", "Spatialiser", "Verbaliser"),
  contrasts = c("_visualiser", "_spatialiser", "_verbaliser"),
  base = 1
)
}
\arguments{
\item{type}{The type of data to return: "experiment", "survey", "clustering",
or "all".}

\item{names}{A character vector of names for the clusters passed to
\code{\link[=add_named_clusters]{add_named_clusters()}}. Default is
\code{c("Spatialiser", "Visualiser", "Verbaliser")}.}

\item{levels}{A character vector of levels for the factor passed to
\code{\link[=add_named_clusters]{add_named_clusters()}}. Default is
\code{c("Visualiser", "Spatialiser", "Verbaliser")}.}

\item{contrasts}{A character vector of contrasts for the factor levels
passed to \code{\link[=add_named_clusters]{add_named_clusters()}}.}

\item{base}{An integer indicating the base level for the contrasts. Default
is 1, which corresponds to the first cluster in \code{names}.}
}
\value{
A cleaned and clustered data frame or a list of cleaned and clustered data
frames, depending on the \code{type} parameter:
\itemize{
\item If \code{type} is "experiment", returns the cleaned experiment data frame with
cluster information.
\item If \code{type} is "survey", returns the cleaned survey data frame with cluster
information.
\item If \code{type} is "clustering", returns the clustering results.
\item If \code{type} is "all", returns a list containing the cleaned experiment data
frame with cluster information, the cleaned survey data frame with cluster
information, and the clustering results.
}
}
\description{
This function retrieves clean data using \code{\link[=get_clean_data]{get_clean_data()}} and then
performs clustering on the OSIVQ data using \code{\link[=cluster_osivq]{cluster_osivq()}}. It then
adds named clusters to the survey data using \code{\link[=add_named_clusters]{add_named_clusters()}} and
merges the cluster information with the experiment data. Depending on the
\code{type} parameter, it returns either the cleaned experiment data with cluster
information, the cleaned survey data with cluster information, the clustering
results, or all of these as a list. This is a convenience function to
streamline the process of obtaining clustered data for analysis, but its
defaults are based on a first iteration of this clustering procedure, which
is described in full in the vignette \code{vignette("osivq_clusters")}.
}
\examples{
colnames(get_clustered_data(type = "experiment"))
}
