% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tools.R
\name{get_params}
\alias{get_params}
\title{Get the fixed parameters of a model in a clean format}
\usage{
get_params(model, ...)
}
\arguments{
\item{model}{A fitted model object.}

\item{...}{Additional arguments passed to \code{parameters::model_parameters()}.}
}
\value{
A formatted data frame with the fixed parameters of the model.
}
\description{
Get the fixed parameters of a model in a clean format
}
\examples{
df_expe <- get_clean_data()$df_expe

if (require("glmmTMB", quietly = TRUE)) {
 model <- glmmTMB::glmmTMB(
  formula = accuracy ~ group_2 * category + (1 | id),
  data = df_expe,
  family = binomial(link = "logit"),
  prior = set_ranef_prior(65)
 )

 get_params(model)
}
}
