% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tools.R
\name{get_singularity}
\alias{get_singularity}
\title{Check if the model is singular and print a message}
\usage{
get_singularity(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
Nothing. Prints a message indicating whether the model is singular
or not.
}
\description{
Check if the model is singular and print a message
}
\examples{
df_expe <- get_clean_data()$df_expe

if (require("glmmTMB", quietly = TRUE)) {
 model <- glmmTMB::glmmTMB(
  formula = accuracy ~ group_2 * category + (1 | id),
  data = df_expe,
  family = binomial(link = "logit"),
  prior = set_ranef_prior(20)
 )

 get_singularity(model)
}
}
