% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_trials.R
\name{get_trial_count}
\alias{get_trial_count}
\title{Get the count of trials per participant and category}
\usage{
get_trial_count(df)
}
\arguments{
\item{df}{A data frame containing trial data with columns for id, Group, and
category.}
}
\value{
A data frame summarizing the number of trials per participant and
category, including total trials and a flag for low trial counts.
}
\description{
Get the count of trials per participant and category
}
