% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_viie_data}
\alias{get_viie_data}
\title{Get data with the Visual Imagery Impedance Effect manually calculated}
\usage{
get_viie_data(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=get_clustered_data]{get_clustered_data()}}.}
}
\value{
A data frame with one row per participant containing their
questionnaire scores and groups along with three VIIE scores:
\code{viie_total} (mean visual RT minus mean non-visual RT), \code{viie_spatial}
(visual RT minus spatial RT), and \code{viie_control} (visual RT minus control
RT).
}
\description{
Another approach to our main problem (evaluating the individual differences
in contrasts between categories for, e.g., response times) could be to
compute the category contrasts in the outcome variable manually and then
compare individuals or groups on these difference scores. This is less
powerful than using mixed-effects models with interaction terms, but some
might wonder what comes out of such an approach. This function computes
these difference scores between response times in the visual category minus
the other categories (spatial, control, and non-visual, an average of
control and spatial) to create three Visual Imagery Impedance Effect (VIIE)
scores per participant. It returns a data frame with these VIIE scores along
with relevant individual difference (questionnaire) variables.
}
\examples{
colnames(get_viie_data())
}
