% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nl_contrasts}
\alias{nl_contrasts}
\title{Results of the contrast analyses performed on the \code{mgcv} non-linear models}
\format{
An object of class \code{list} of length 3.
}
\usage{
nl_contrasts
}
\description{
These analyses were saved as package data to compensate for the fact that
the non-linear model fits themselves were too heavy to be saved in the
package. They are used to display the statistics in the
\code{vignette("nl_modelling")}.
}
\keyword{datasets}
