% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nl_predictions}
\alias{nl_predictions}
\title{Model predictions computed with the \code{mgcv} non-linear model fits}
\format{
An object of class \code{list} of length 3.
}
\usage{
nl_predictions
}
\description{
These predictions were saved as package data to compensate for the fact that
the non-linear model fits themselves were too heavy to be saved in the
package. They are used to visualise the non-linear dynamics in the
\code{vignette("nl_modelling")}.
}
\keyword{datasets}
