% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivoting_variables.R
\name{pivot_terms_longer}
\alias{pivot_terms_longer}
\title{Get a long format data frame with the problem terms in a single column}
\usage{
pivot_terms_longer(df)
}
\arguments{
\item{df}{A data frame containing the problem terms with columns for
\code{premise_1_rt}, \code{premise_2_rt}, \code{premise_3_rt}, and \code{conclusion_rt} along
with other relevant columns like \code{group}, \code{category}, and \code{problem}.}
}
\value{
A data frame in long format with columns with the problem term a an
integer in a \code{term} column and as a factor in a \code{term_name} column.
}
\description{
Get a long format data frame with the problem terms in a single column
}
\examples{
get_clean_data()$df_expe |>
  filter_trials_on_rt() |>
  pivot_terms_longer() |>
  dplyr::select(id, problem, category, term, term_name, rt)
}
