% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_others.R
\name{plot_nl}
\alias{plot_nl}
\title{Plot model predictions from the non-linear GAMM results}
\usage{
plot_nl(df, title = NULL, ...)
}
\arguments{
\item{df}{A data frame containing the model predictions with columns:
\itemize{
\item \code{term}: the trial term as an integer from 1 to 4
\item \code{Predicted}: the predicted response time
\item \code{CI_low}: the lower bound of the confidence interval
\item \code{CI_high}: the upper bound of the confidence interval
\item \code{category}: the problem category (e.g., "Visual", "Control", "Spatial")
\item \code{group}: the participant group
Results from model fits are saved natively in the package in the
\code{nl_predictions} object.
}}

\item{title}{An optional title for the plot. Default is NULL.}

\item{...}{Additional arguments passed to the \code{\link[=theme_pdf]{theme_pdf()}} function for
customising the plot theme.}
}
\value{
A ggplot2 object showing the model predictions with confidence
intervals, faceted by participant group and coloured by problem category.
}
\description{
Plot model predictions from the non-linear GAMM results
}
\examples{
plot_nl(nl_predictions$vviq_2, base_size = 12)
}
