% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating_data_analyses.R
\name{plot_power}
\alias{plot_power}
\title{Plot the results of a power analysis by simulation}
\usage{
plot_power(power_results, threshold = 0.05, ...)
}
\arguments{
\item{power_results}{A data frame with the results of the power analysis by
simulation.}

\item{threshold}{The significance threshold to use for calculating power.
Default is 0.05.}

\item{...}{Additional arguments passed to the \code{theme_pdf()} function for
customizing the plot theme.}
}
\value{
A ggplot2 object with the power curves.
}
\description{
This function plots the results of a power analysis by simulation. It takes
a data frame similar to those output by the \code{run_power_analysis()} function.
}
\examples{
plot_power(power_results) # power_results is natively in the package
}
