% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power_results}
\alias{power_results}
\title{Results of the power analysis by simulation presented in the manuscript}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 147000 rows and 4 columns.
}
\usage{
power_results
}
\description{
This dataset contains the results of the power analysis conducted by
fitting models on 147,000 simulated datasets with various sample sizes
and effect sizes. The results are used to estimate the statistical power
of the study to detect effects of different sizes with different sample
sizes. This analysis took 16 hours and 44 minutes on a machine with 40 cores,
so naturally the dataset is included here to avoid having to redo it. It
contains the same columns as the results of the \code{run_power_analysis()}:
\itemize{
\item \code{n_subj_per_group} (number of subjects per group in the simulated dataset)
\item \code{beta_vis} (the effect size used in the simulation, in seconds)
\item \code{n_simulations} (the identifier of the simulation, from 1 to the number
of simulations per combination, here 350)
\item \code{p_value} (the p-value of the effect of interest in the fitted model)
}
}
\keyword{datasets}
