% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_with_superb.R
\name{prepare_df_for_plotting}
\alias{prepare_df_for_plotting}
\title{Helper function to prepare the data frame for plotting}
\usage{
prepare_df_for_plotting(df, dvar)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{dvar}{The dependent variable to be averaged and plotted.}
}
\value{
A data frame ready for plotting with the mean dependent variable
for each participant, category, and grouping/cluster combination.
}
\description{
This function groups the data by participant ID, category, and any grouping
or clustering variables, then calculates the mean of the dependent variable
for each group. It also reverses the factor levels for any clustering
variables to ensure correct ordering in plots.
}
\examples{
df <- get_clean_data()$df_expe
df_rt <- filter_trials_on_rt(df)

df_rt |> prepare_df_for_plotting(dvar = rt_total) |> head(12)

}
\keyword{internal}
