% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tools.R
\name{report_contrast}
\alias{report_contrast}
\title{Get the contrasts of a model and format them for reporting}
\usage{
report_contrast(model, formula, ...)
}
\arguments{
\item{model}{A fitted model object.}

\item{formula}{A formula specifying the variables for which to get the
contrasts. See \code{?emmeans::emmeans} for details.}

\item{...}{Additional arguments passed to \code{get_contrast()}.}
}
\value{
A data frame with the pairwise contrasts of the specified variables
formatted for reporting.
}
\description{
Get the contrasts of a model and format them for reporting
}
\examples{
df_expe <- get_clean_data()$df_expe

if (require("glmmTMB", quietly = TRUE)) {
 model <- glmmTMB::glmmTMB(
  formula = accuracy ~ group_2 * category + (1 | id),
  data = df_expe,
  family = binomial(link = "logit"),
  prior = set_ranef_prior(20)
 )

 report_contrast(model, ~ category | group_2)
}
}
