% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tools.R
\name{report_rope}
\alias{report_rope}
\title{Report the ROPE analysis for marginal effects}
\usage{
report_rope(marg_effects, ..., digits = 2)
}
\arguments{
\item{marg_effects}{A marginaleffects object obtained with
\code{\link[marginaleffects:comparisons]{marginaleffects::avg_comparisons()}} containing the contrasts to analyse.}

\item{...}{Grouping variables for summarising the ROPE results.}

\item{digits}{Number of decimal places to round the results. Default is 3.}
}
\value{
A data frame summarising the ROPE analysis with the estimates,
95\% CIs, and proportions of draws within, below, and above the ROPE.
}
\description{
Report the ROPE analysis for marginal effects
}
