% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{Custom ggsave wrapper set with Nature's formatting guidelines}
\usage{
save_plot(
  plot,
  path,
  ncol = 1,
  height = 90,
  width = NULL,
  print_it = FALSE,
  verbose = TRUE,
  units = "mm",
  dpi = 600,
  ...
)
}
\arguments{
\item{plot}{The plot to save.}

\item{path}{The path to save the plot.}

\item{ncol}{The number of columns for the plot. Either 1 (default) or 2.}

\item{height}{The height of the plot in mm. Default is 90 mm.}

\item{width}{Optional. The width of the plot in mm. If NULL (default), it
will be set to 88 mm for one-column figures and 180 mm for
two-column figures.}

\item{print_it}{Logical. Whether to print the plot in the console. Default is
FALSE.}

\item{verbose}{Logical. Whether to print a message in the console. Default
is TRUE.}

\item{units}{The units for the width and height. Default is "mm".}

\item{dpi}{The resolution of the plot. Default is 600.}

\item{...}{Additional arguments passed to \code{ggsave()}.}
}
\value{
Nothing. The function saves the plot to the specified path.
}
\description{
See:
https://www.nature.com/documents/NRJs-guide-to-preparing-final-artwork.pdf
It's pretty strict. A one column figure is 88 mm wide and a two column figure
is 180 mm wide. Depending on the length of the figure caption, there are
different maximum heights (see the PDF). Most figures types must be in vector
format to prevent quality loss when zooming in. Ever since I found these
guidelines, I use them for all figures, even if they are not for Nature...
Because it looks nice and I like it.
}
