% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tools.R
\name{set_ranef_prior}
\alias{set_ranef_prior}
\title{Create a weakly informative regularizing Gamma prior for the random effects}
\usage{
set_ranef_prior(gamma_mean = 100)
}
\arguments{
\item{gamma_mean}{A numeric value indicating the mean of the Gamma prior. The
higher the value, the less impact the prior has, but also the more chance of
singularity.}
}
\value{
A data frame with the prior and class for the random effects.
}
\description{
Create a weakly informative regularizing Gamma prior for the random effects
}
\examples{
set_ranef_prior(100)
}
