% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating_data_analyses.R
\name{simulate_acc_data}
\alias{simulate_acc_data}
\title{Simulate accuracy data for the factorial design}
\usage{
simulate_acc_data(
  n_subj_per_group,
  beta_0 = 1.36,
  tau_0 = 0.05,
  tau_vis = 0.01,
  tau_spa = 0.01,
  beta_aph = 0,
  beta_vis = 0,
  beta_spa = 0,
  beta_aph_vis = 0,
  beta_aph_spa = 0,
  seed = NULL,
  ...
)
}
\arguments{
\item{n_subj_per_group}{The number of subjects per group.}

\item{beta_0}{The intercept for the model.}

\item{tau_0}{The standard deviation of the random intercept for each
subject.}

\item{tau_vis}{The standard deviation of the random slope for the visual
category.}

\item{tau_spa}{The standard deviation of the random slope for the spatial
category.}

\item{beta_aph}{The fixed effect of the aphantasia group.}

\item{beta_vis}{The fixed effect of the visual category.}

\item{beta_spa}{The fixed effect of the spatial category.}

\item{beta_aph_vis}{The interaction between the aphantasia group and the
visual category.}

\item{beta_aph_spa}{The interaction between the aphantasia group and the
spatial category.}

\item{seed}{The seed for random number generation. If NULL, the seed is not
set.}

\item{...}{Additional arguments passed to the function. Unused.}
}
\value{
A data frame with the simulated accuracy data.
}
\description{
This function simulates accuracy data for the factorial design of the
experiment.

It was designed for potential power analyses on accuracy data, but
it ended up being unused due to a lack of previous data to base the analyses
on. The power analyses were conducted on RT data instead.

It creates a data frame with the following columns:
\itemize{
\item \code{id}: A unique identifier for each subject.
\item \code{Group}: The group to which the subject belongs (aphantasia or typical).
\item \code{category}: The category of the trial (visual, spatial, or control).
\item \code{trial}: The trial number (1 to 9).
\item \code{accuracy}: The accuracy of the subject's response (1 for correct, 0 for
incorrect).
Inspired by \href{https://debruine.github.io/lmem_sim/articles/appendix3a_binomial.html}{Lisa DeBruine}.
}
}
\examples{
df <- simulate_acc_data(100)
head(df)

df |>
  dplyr::group_by(group, category) |>
  dplyr::reframe(
    mean_acc = mean(accuracy),
    median_acc = median(accuracy),
    sd_acc = sd(accuracy)
    )
}
\keyword{internal}
