% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating_data_analyses.R
\name{simulate_rt_data}
\alias{simulate_rt_data}
\title{Simulate skewed RT data for the factorial design}
\usage{
simulate_rt_data(
  n_subj_per_group,
  meanlog = 2.1,
  sdlog = 0.55,
  shift = 5,
  tau_0 = 0.9,
  tau_vis = 0.75,
  tau_spa = 0,
  beta_aph = 0,
  beta_vis = 0,
  beta_spa = 0,
  beta_aph_vis = 0,
  beta_aph_spa = 0,
  seed = NULL,
  ...
)
}
\arguments{
\item{n_subj_per_group}{Number of subjects per group. The total number of
subjects will be \code{n_subj_per_group * 2}.}

\item{meanlog}{Mean of the log-normal distribution for the base RT values.}

\item{sdlog}{Standard deviation of the log-normal distribution for the base
RT values.}

\item{shift}{Non-decision time, i.e., the minimum value of the RTs.}

\item{tau_0}{By-subject random intercept standard deviation.}

\item{tau_vis}{By-subject random slope standard deviation for the visual
category.}

\item{tau_spa}{By-subject random slope standard deviation for the spatial
category.}

\item{beta_aph}{Effect of the aphantasia group on the RTs.}

\item{beta_vis}{Effect of the visual category on the RTs.}

\item{beta_spa}{Effect of the spatial category on the RTs.}

\item{beta_aph_vis}{Effect of the interaction between the aphantasia group
and the visual category on the RTs.}

\item{beta_aph_spa}{Effect of the interaction between the aphantasia group
and the spatial category on the RTs.}

\item{seed}{Random seed for reproducibility. If \code{NULL}, no seed is set.}

\item{...}{Additional arguments passed to the function. Unused.}
}
\value{
A dataframe with the simulated RT data.
}
\description{
This function simulates response time data for the factorial design of the
experiment. It was used for power analyses by simulation.

Inspired by \href{https://cjungerius.github.io/powersim/}{Chris Jungerius}.
}
\examples{
# No main effects
df <- simulate_rt_data(100)
head(df)

df |>
  dplyr::group_by(group, category) |>
  dplyr::reframe(
    mean_rt = mean(rt_total),
    median_rt = median(rt_total),
    sd_rt = sd(rt_total),
    min_rt = min(rt_total),
    max_rt = max(rt_total)
  )

# Visual category effect around 2.5s
df <- simulate_rt_data(100, beta_vis = 2.5)
df |>
  dplyr::group_by(group, category) |>
  dplyr::reframe(
    mean_rt = mean(rt_total),
    median_rt = median(rt_total),
    sd_rt = sd(rt_total),
    min_rt = min(rt_total),
    max_rt = max(rt_total)
  )
}
