% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating_data_analyses.R
\name{simulate_rt_test}
\alias{simulate_rt_test}
\title{Simulate data, fit a model on it and test our hypothesis}
\usage{
simulate_rt_test(n_subj_per_group, beta_vis, p = progressr::progressor(), ...)
}
\arguments{
\item{n_subj_per_group}{Number of subjects per group. The total number of
subjects will be \code{n_subj_per_group * 2}.}

\item{beta_vis}{Effect of the visual category on the RTs. This is used to
create the main effect of interest by adding this effect to the typical
group only (thus creating an interaction between group and category).}

\item{p}{A progressor object from the \code{progressr} package to update the
progress bar during the simulation.}

\item{...}{Additional arguments passed to the \code{simulate_rt_data} function.
These can include parameters like \code{meanlog}, \code{sdlog}, \code{shift}, etc.}
}
\value{
The p-value for the group x control-visual interaction contrast.
}
\description{
This function simulates data based on the parameters provided, fits a model
to the simulated data, and extracts the p-value for the group x
control-visual interaction contrast (our theoretical contrast of interest).
This function is used repeatedly in a power analysis by simulation.
}
\examples{
simulate_rt_test(10, 1)   # Low power, unlikely to detect the effect
simulate_rt_test(40, 1.5) # Better power, more likely
simulate_rt_test(150, 1)  # High power, highly likely even for a small effect
}
