% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating_data_analyses.R
\name{simulate_strats_data}
\alias{simulate_strats_data}
\title{Simulate strategies data for all four groups}
\usage{
simulate_strats_data(
  n,
  means_aph = c(-2, 1, -1.3, -1.6, -0.7),
  means_hypo = c(-1.1, 0.1, -1.2, -1.6, -0.1),
  means_typical = c(0.1, 0.1, -1.5, -1.4, -0.3),
  means_hyper = c(-0.1, -0.2, -1.7, -1.7, -0.9),
  sd = 0.5,
  modelling_version = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{n}{Number of participants per group.}

\item{means_aph}{Numeric vector of means for the Aphantasia group.}

\item{means_hypo}{Numeric vector of means for the Hypophantasia group.}

\item{means_typical}{Numeric vector of means for the Typical group.}

\item{means_hyper}{Numeric vector of means for the Hyperphantasia group.}

\item{sd}{Standard deviation for all groups (default is 0.5).}

\item{modelling_version}{Logical indicating whether to return the data in a
format suitable for modelling with \code{faux::add_contrast()}. If \code{TRUE}, the
data will be returned with contrasts added for the groups and the
variables will be converted to ordered factors. If \code{FALSE}, the data will be
returned in its original format with the group names as character strings.}

\item{seed}{Optional seed for reproducibility of the random data generation.}

\item{...}{Additional arguments passed to the function. Unused.}
}
\value{
A tibble containing the simulated strategies data for the four
groups.
}
\description{
This function simulates strategies data for the four groups of phantasia
based on the means and standard deviations provided. The data is simulated
using the \code{latent2likert::rlikert()} function, which generates
likert-type data based on latent variables. The function allows for
customization of the means and standard deviations for each group, as well as
the option to generate data in a format suitable for modelling with
\code{faux::add_contrast()}. The default means and standard deviations are based
on the findings of Reeder & Pounder (2024) on strategies used by
people with different levels of imagery.

It was designed for potential power analyses on strategies data, but
it ended up being unused due to a lack of time and previous data to build
upon. The power analyses were conducted on RT data instead.
}
\examples{
df <- simulate_strats_data(n = 100, seed = 123)
head(df)

df |>
 dplyr::group_by(group) |>
 dplyr::reframe(dplyr::across(tidyselect::contains("_strat"), mean))
}
\keyword{internal}
