% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_osivq.R
\name{add_named_clusters}
\alias{add_named_clusters}
\title{Add a column with named cluster assignments to a data frame}
\usage{
add_named_clusters(
  df,
  clustering,
  method = "CSPA",
  names = c("cluster_1", "cluster_2", "cluster_3"),
  levels = names,
  contrasts = paste0("_", levels),
  base = 1,
  ...
)
}
\arguments{
\item{df}{A data frame with the OSIVQ scores, typically obtained from
\code{\link[=get_clean_data]{get_clean_data()}}.}

\item{clustering}{A clustering object obtained from \code{\link[=cluster_osivq]{cluster_osivq()}}.}

\item{method}{A character string specifying the consensus clustering method
to use. Must be "kmodes", "majority", or "CSPA". Default is "CSPA".}

\item{names}{A character vector of names for the clusters. Default is
\code{c("cluster_1", "cluster_2", "cluster_3")}.}

\item{levels}{A character vector of levels for the factor. Default is the
same as \code{names}.}

\item{contrasts}{A character vector of contrasts for the factor levels.}

\item{base}{An integer indicating the base level for the contrasts. Default
is 1, which corresponds to the first cluster in \code{names}.}

\item{...}{Additional arguments passed to \code{\link[=add_factor_contrasts]{add_factor_contrasts()}}.}
}
\value{
A data frame with an additional column \code{cluster} that contains the
named cluster assignments.
}
\description{
This function is tailored for three clusters. After checking the results of
the clustering procedure with \code{\link[=summarise_clustering]{summarise_clustering()}} and a graphical
examination, we can choose names for the three clusters manually and input
them into this function to get a meaningful variable to analyse.
}
\examples{
df <- get_clean_data()$df_survey
clustering <- cluster_osivq(df)

df |> add_named_clusters(clustering) |> dplyr::select(id, group_4, cluster)
}
