% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_osivq.R
\name{cluster_osivq}
\alias{cluster_osivq}
\title{Cluster the OSIVQ data using consensus between various algorithms}
\usage{
cluster_osivq(
  df,
  algorithms = c("gmm", "pam", "cmeans"),
  cons.funs = c("kmodes", "majority", "CSPA"),
  seed = 667,
  progress = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{A dataframe with the OSIVQ scores, typically obtained from
\code{\link[=get_clean_data]{get_clean_data()}}.}

\item{algorithms}{A character vector of clustering algorithms to use. Default
is \code{c("gmm", "pam", "cmeans", "diana")}. See \code{?diceR::dice()} for more
details.}

\item{cons.funs}{A character vector of consensus functions to use. Default is
\code{c("CSPA")}. See \code{?diceR::dice()} for more details.}

\item{seed}{An integer seed for reproducibility. Default is 667.}

\item{progress}{Logical value indicating whether to show a progress bar.}

\item{verbose}{Logical value indicating whether to print detailed messages
during the clustering process.}
}
\value{
A list with the clustering results from \code{\link[diceR:dice]{diceR::dice()}}.
}
\description{
Cluster the OSIVQ data using consensus between various algorithms
}
