% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experiment_data}
\alias{experiment_data}
\title{Reasoning experiment data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3672 rows and 15 columns.
}
\source{
Data collected through an online experiment
}
\usage{
experiment_data
}
\description{
This dataset contains the results of the main reasoning experiment. It
contains the following columns:
\itemize{
\item An encrypted unique participant \code{id} (that also allows to join the data
with the survey data)
\item \code{language} (the language the experiment was conducted in, FR or EN)
\item \code{group} (aphantasia, hypophantasia, typical, hyperphantasia or no_vviq)
\item \code{expe_phase} (the phase of the experiment)
\item \code{trial_number} (the presentation order of the problems)
\item \code{problem} (the problem presented to the participant)
\item \code{category} (the category of the problem, visual, spatial or control)
\item \verb{premise_*_rt} (the reaction time of the participant for each premise)
\item \code{conclusion_rt} (the reaction time of the participant for the conclusion,
i.e., the response screen)
\item \code{rt_total} (the total time spent on the trial)
\item \code{response} (the participant's response, TRUE or FALSE)
\item \code{correct_response} (the correct response to the problem, TRUE or FALSE)
\item \code{accuracy} (the accuracy of the participant's response, TRUE or FALSE)
}
}
\keyword{datasets}
