% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tools.R
\name{get_contrast}
\alias{get_contrast}
\title{Get the pairwise contrasts of variables in a model}
\usage{
get_contrast(model, formula, at = NULL, ...)
}
\arguments{
\item{model}{A fitted model object.}

\item{formula}{A formula specifying the variables for which to get the
contrasts. See \code{?emmeans::emmeans} for details.}

\item{at}{Optional. A list of values at which to evaluate the contrasts.}

\item{...}{Additional arguments passed to \code{emmeans::contrast()}.}
}
\value{
An emm_grid object with the pairwise contrasts of the specified
variables.
}
\description{
Get the pairwise contrasts of variables in a model
}
