% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_utils.R
\name{get_params}
\alias{get_params}
\title{Get the fixed parameters of a model in a clean format}
\usage{
get_params(model, ...)
}
\arguments{
\item{model}{A fitted model object.}

\item{...}{Additional arguments passed to \code{parameters::model_parameters()}.}
}
\value{
A formatted data frame with the fixed parameters of the model.
}
\description{
Get the fixed parameters of a model in a clean format
}
