% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tools.R
\name{get_performance}
\alias{get_performance}
\title{Get performance indices for a model in a clean format}
\usage{
get_performance(model, ...)
}
\arguments{
\item{model}{A fitted model object.}

\item{...}{Additional arguments passed to \code{performance::model_performance()}.}
}
\value{
A formatted data frame with the performance indices of the model.
}
\description{
Get performance indices for a model in a clean format
}
