% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_utils.R
\name{get_singularity}
\alias{get_singularity}
\title{Check if the model is singular and print a message}
\usage{
get_singularity(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
Nothing. Prints a message indicating whether the model is singular
or not.
}
\description{
Check if the model is singular and print a message
}
