% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivoting_variables.R
\name{pivot_strategies_longer}
\alias{pivot_strategies_longer}
\title{Get a long format data frame with the strategies gathered in a single column}
\usage{
pivot_strategies_longer(df, base = 1, ordered = FALSE, ...)
}
\arguments{
\item{df}{A data frame containing the strategies data with columns for
\code{id}, \code{group}, \code{visual_strat}, \code{spatial_strat}, \code{verbal_strat},
\code{semantic_strat}, and \code{sensorimotor_strat}.}

\item{base}{An integer indicating the base level for factor contrasts.
Default is 1 for the visual strategy.}

\item{ordered}{A logical indicating whether the \code{score} factor levels should
be ordered. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[=add_factor_contrasts]{add_factor_contrasts()}}.}
}
\value{
A data frame in long format with columns for \code{id}, \code{group},
\code{strategy}, and \code{score}, where \code{strategy} is a factor with levels for each
strategy and \code{score} is a factor indicating the level of strategy use.
}
\description{
Get a long format data frame with the strategies gathered in a single column
}
\examples{
get_clean_data("survey") |>
  pivot_strategies_longer() |>
  dplyr::select(id, group_4, strategy, score)
}
