% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_median_rt_distribution.R
\name{plot_median_rt_distribution}
\alias{plot_median_rt_distribution}
\title{Plot the distribution of the median RT across participants}
\usage{
plot_median_rt_distribution(df, sd_mult = 2, ...)
}
\arguments{
\item{df}{A data frame containing participant responses with an \code{id} column
and a \code{rt_total} for each trial.}

\item{sd_mult}{A numeric value indicating how many standard deviations to use
for identifying suspicious median RTs. The default is 2, which means that
median RTs that are more than 2 standard deviations away from the mean
will be considered suspicious.}

\item{...}{Additional arguments passed to the \code{\link[=theme_pdf]{theme_pdf()}} function for
customizing the plot theme.}
}
\value{
A ggplot2 object showing the distribution of median RTs across
participants, with suspicious median RTs highlighted.
}
\description{
Plot the distribution of the median RT across participants
}
\examples{
df <- get_clean_data(sd_mult = 10)$df_expe
plot_median_rt_distribution(df)
}
