% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_osivq_ternary.R
\name{plot_osivq_ternary}
\alias{plot_osivq_ternary}
\title{Plot the OSIVQ scores of clusters in a ternary diagram}
\usage{
plot_osivq_ternary(
  df,
  dot_size = 1.5,
  plot_it = FALSE,
  colours = palette.colors(),
  ...
)
}
\arguments{
\item{df}{A dataframe containing \code{osivq_object}, \code{osivq_spatial},
\code{osivq_verbal} and \code{cluster} columns.}

\item{dot_size}{Size of the dots in the plot.}

\item{plot_it}{Logical. If TRUE, the plot will be printed immediately.}

\item{colours}{A vector of colours to use for the clusters in the plot passed
to the \code{\link[ggplot2:scale_manual]{ggplot2::scale_discrete_manual()}} function. Default is
\code{\link[=palette.colors]{palette.colors()}}.}

\item{...}{Additional arguments passed to the \code{\link[=theme_pdf]{theme_pdf()}} function.}
}
\value{
A ggplot object representing the OSIVQ scores in a ternary diagram.
}
\description{
Plot the OSIVQ scores of clusters in a ternary diagram
}
