% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strategies.R
\name{plot_strategies_barplot}
\alias{plot_strategies_barplot}
\title{Plot proportions of strategy use for groups as barplots}
\usage{
plot_strategies_barplot(df_long, grouping = group, ...)
}
\arguments{
\item{df_long}{A data frame containing the strategies data in a long format
with columns for grouping, strategy, and score.}

\item{grouping}{A variable to group the data by, typically group or cluster.
Can be unquoted.}

\item{...}{Additional arguments passed to the \code{\link[=theme_pdf]{theme_pdf()}} function for
customizing the plot theme.}
}
\value{
A ggplot2 object showing the proportions of strategy use for each
group as a barplot, with strategies as facets.
}
\description{
Plot proportions of strategy use for groups as barplots
}
