% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strategies.R
\name{plot_strategies_scores}
\alias{plot_strategies_scores}
\title{Plot mean strategy scores for groups}
\usage{
plot_strategies_scores(
  df_long,
  grouping = "group",
  title = NULL,
  x_labels = NULL,
  ...
)
}
\arguments{
\item{df_long}{A data frame containing the strategies data in a long format
with columns for grouping, strategy, and score.}

\item{grouping}{A variable to group the data by, typically group or cluster.
Can be unquoted.}

\item{title}{Optional title for the plot.}

\item{x_labels}{Optional labels for the x-axis, if different from the
grouping variable.}

\item{...}{Additional arguments passed to the \code{\link[=theme_pdf]{theme_pdf()}} function for
customizing the plot theme.}
}
\value{
A ggplot2 object showing the mean strategy scores for each group,
with strategies as colors and filled points, and error bars representing
standard error of the mean.
}
\description{
Plot mean strategy scores for groups
}
\examples{
df_long <-
  get_clean_data()$df_survey |>
  pivot_strategies_longer()

plot_strategies_scores(df_long, grouping = group, base_size = 12)
}
