% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_osivq.R
\name{summarise_clustering}
\alias{summarise_clustering}
\title{Get the size and questionnaire means of clusters}
\usage{
summarise_clustering(df)
}
\arguments{
\item{df}{A data frame with columns \code{group}, \code{cluster}, \code{vviq_total_score},
\code{osivq_object}, \code{osivq_spatial}, \code{osivq_verbal}, and \code{raven_score}.}
}
\value{
A data frame summarizing the clusters, including the number of
participants in each cluster and the mean scores for VVIQ, OSIVQ object,
OSIVQ spatial, OSIVQ verbal, and Raven scores.
}
\description{
Get the size and questionnaire means of clusters
}
