% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{Participant's questionnaire data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 125 rows and 105 columns.
}
\source{
Data collected through online experiments
}
\usage{
survey_data
}
\description{
This dataset contains the responses to the questionnaires collected from the
participants. It has been gathered through several online experiments that
ran simulatneously on the
\href{https://innerexperiencelab.com}{innerexperiencelab.com} website. It
contains the following:
\itemize{
\item An encrypted unique participant \code{id} (that also allows to join the data
with the experiment data)
\item The \code{language} the experiment was conducted in (FR or EN)
\item \code{age}
\item \code{gender}
\item \code{Group} (aphantasia, hypophantasia, typical, hyperphantasia or no_vviq)
\item \code{country}
\item \code{language_native} (the native language of the participant)
\item \code{language_usual} (the language the participant usually uses)
\item \code{job} (the job of the participant)
\item \code{vviq_is_complete} (whether the participant completed the VVIQ)
\item \code{vviq_total_score} (the total score of the VVIQ)
\item All the VVIQ items
\item \code{osivq_is_complete} (whether the participant completed the OSIVQ)
\item \code{osivq_object}, \code{osivq_spatial}, \code{osivq_verbal} (the mean scores of the
OSIVQ subscales)
\item All the OSIVQ items
\item \code{raven_is_complete} (whether the participant completed the Raven's
Progressive Matrices)
\item \code{raven_score} (the score of the Raven's Progressive Matrices)
\item \code{nieq_is_complete} (whether the participant completed the NIEQ)
\item All the NIEQ items on the two "frequency" and "proportion" sub-scales
\item \verb{*_false_info} (self-report of whether the participant provided false
information on the questionnaires, people are usually honest about this)
\item \code{met_issues} (self-report of whether the participant encountered issues
during the experiment)
\item \verb{*_external_support} (self-report of whether the participant used
external support during the experiment, e.g. paper)
\item \verb{strats_*} (self-report of the strategies used by the participant during
the experiment)
\item \code{has_adhd} (self-report of whether the participant has ADHD)
\item \code{has_asd} (self-report of whether the participant has ASD)
\item \code{has_dyslexia} (self-report of whether the participant has dyslexia)
\item \code{has_other_neuro_trouble} (self-report of whether the participant has
other neurological troubles)
\item \code{has_treatment} (self-report of whether the participant is under treatment
for a neurological trouble)
\item \code{has_been_distracted} (self-report of whether the participant has been
distracted during the experiment)
\item \code{has_cheated} (self-report of whether the participant has used external
help to answer the questions)
}
}
\details{
Since the data has been collected in different experiments, not all
participants have NIEQ data, which was not central for this specific
paradigm. It has been included here for exploratory purposes.
}
\keyword{datasets}
